+++ src/main/java/com/example/post/service/CommentService.java
@@ -1,9 +1,9 @@
 package com.example.post.service;
 
 import com.example.post.entity.Board;
-import com.example.post.entity.BoardHeart;
 import com.example.post.entity.Comment;
 import com.example.post.entity.CommentHeart;
+import com.example.post.jwt.JwtUtil;
 import com.example.post.repository.BoardRepository;
 import com.example.post.repository.CommentHeartRepository;
 import com.example.post.repository.CommentRepository;
@@ -20,6 +20,7 @@ public class CommentService {
     private final CommentRepository commentRepository;
     private final BoardRepository boardRepository;
     private final CommentHeartRepository commentHeartRepository;
+    private final JwtUtil jwtUtil;
 
     public ResponseEntity<?> write(String content, String email, long boardId) {
          Board board = boardRepository.findById(boardId).orElse(null);
@@ -37,11 +38,21 @@ public class CommentService {
          return ResponseEntity.ok("댓글이 작성되었습니다.");
     }
 
-    public ResponseEntity<?> update(long cmtId, String content) {
+    public ResponseEntity<?> update(String token, long cmtId, String content) {
         Comment comment = commentRepository.findById(cmtId).orElse(null);
 
         if(comment == null)
             return ResponseEntity.badRequest().body("댓글이 존재하지 않습니다.");
+
+        String email;
+        try{
+            email = jwtUtil.getEmail(token);
+        }catch (Exception e){
+            return ResponseEntity.badRequest().body("유효하지 않은 토큰");
+        }
+
+        if(!email.equals(comment.getEmail()))
+            return ResponseEntity.badRequest().body("작성자만 수정할 수 있습니다.");
 
         comment.setContent(content);
         commentRepository.save(comment);
@@ -50,7 +61,23 @@ public class CommentService {
     }
 
     @Transactional
-    public ResponseEntity<?> delete(long cmtId) {
+    public ResponseEntity<?> delete(String token, long cmtId) {
+        Comment comment = commentRepository.findById(cmtId).orElse(null);
+
+        if(comment == null)
+            return ResponseEntity.badRequest().body("댓글이 존재하지 않습니다.");
+
+        String email,role;
+        try{
+            email = jwtUtil.getEmail(token);
+            role = jwtUtil.getRole(token);
+        }catch (Exception e){
+            return ResponseEntity.badRequest().body("유효하지 않은 토큰");
+        }
+
+        if(!email.equals(comment.getEmail()) && !role.equals("ADMIN") )
+            return ResponseEntity.badRequest().body("작성자 or 관리자만 삭제할 수 있습니다.");
+
         commentRepository.deleteById(cmtId);
         return ResponseEntity.ok("댓글이 삭제되었습니다.");
     }
@@ -83,7 +110,7 @@ public class CommentService {
         Comment comment = commentRepository.findById(cmtId).orElse(null);
         if(comment == null)
             return ResponseEntity.badRequest().body("존재하지 않는 댓글입니다.");
-        
+
         Optional<CommentHeart> commentHeart = commentHeartRepository.findById(cmtId);
 
         if(!commentHeart.isPresent())
